Refactor build (#23)
* Split the artefacts build

Separate out build for native and container. Container build will need
a different way of handling configuration files.

* Rename init config home dir callback

This is to prep it for conditional compilation.

* Fix configuration folder operations

Directly initialise operations variables instead of via init function.
Add more test coverage.

* Update Go mod

* Change container-based

From scratch to sawshell. This is to take advantage of Sawtooth key
generators.

* Update e2e test deployment

This is to reflect the new docker image.

* Change CLI start sequence

It now creates a configuration YAML file.

Signed-off-by: Paul Sitoh <paulwizviz@gmail.com>